requirements.txt
flask
flask-cors
flask-mysqldb
pyjwt
uuid
